feat: implement recovery mechanism for fault tolerance

+++ src/services/heartbeat.ts
@@ -1,5 +1,7 @@
 import * as net from 'net';
 import Node from '../interfaces/node.types';
+import { recoverNode } from './nodeController';
+import commands from '../data/commands';
 
 const HEARTBEAT_INTERVAL = 5000;
 const HEARTBEAT_TIMEOUT = 3500;
@@ -21,25 +23,37 @@ export function startSniffing(nodes: Node[]) {
             console.warn(
               `Heartbeat to server on port ${port} timed out. Marking as DEAD.`
             );
-            node.isRunning = false;
+            node.status = 'DEAD';
             client.destroy();
           }
         }, HEARTBEAT_TIMEOUT);
       });
-      client.on('data', (data) => {
+      client.on('data', async (data) => {
         const response = data.toString().trim();
         clearTimeout(heartbeatTimeout);
         if (response === 'ALIVE') {
-          node.isRunning = true;
-          console.log(`Server on port ${port} is ACTIVE and responding.`);
-        } else {
-          node.isRunning = false;
-          console.error(
-            `Unexpected response from server on port ${port}: ${response}. Marking as DEAD.`
-          );
-        }
+          if (node.status === 'DEAD') {
+            node.status = 'RECOVER';
+            console.log(`Server on port ${port} is RECOVERING...`);
+            try {
+              await recoverNode(Number(port), host, commands);
+              node.status = 'RECOVERED';
+              console.log(`Server on port ${port} is RECOVERED.`);
+            } catch (error) {
+              console.error(`Error during node recovery: ${error}`);
+              node.status = 'RECOVER_FAILED';
+            }
+          } else if (node.status === 'RECOVERED') {
+            node.status = 'ACTIVE';
+            console.log(`Server on port ${port} is ACTIVE and responding...`);
+          } else if (node.status === 'ACTIVE') {
+            console.log(`Server on port ${port} is ACTIVE and responding...`);
+          }
 
-        client.destroy();
+          if (node.status === 'ACTIVE') {
+            client.destroy();
+          }
+        }
       });
       client.on('error', (err) => {
         clearTimeout(heartbeatTimeout);
@@ -47,12 +61,12 @@ export function startSniffing(nodes: Node[]) {
         console.warn(
           `Heartbeat to server on port ${port} timed out. Marking as DEAD.`
         );
-        node.isRunning = false;
+        node.status = "DEAD";
         client.destroy();
       });
       client.on('end', () => {
         console.log(`Connection with server on port ${port} ended.`);
-        node.isRunning = false;
+        node.status = "DEAD";
       });
     });
   }, HEARTBEAT_INTERVAL);
